+++ JsonLibrary/JsonLibrary/ArrayValue.cpp
@@ -9,7 +9,7 @@ ArrayValue::ArrayValue()
 
 ArrayValue::~ArrayValue()
 {
-	for (size_t i = 0; i < size; i++)
+	for (int i = 0; i < size; i++)
 	{
 		delete array->getElementAt(i);
 	}
@@ -24,7 +24,7 @@ void ArrayValue::append(Value* element) {
 std::string ArrayValue::serialize() const
 {
 	std::string output = "[";
-	for (size_t i = 0; i < size; i++)
+	for (int i = 0; i < size; i++)
 	{
 		output.append(this->array->getElementAt(i)->serialize());
 		if (i != size-1)
